Enable clang-diagnostic-shorten-64-to-32
439 sites over all 840 translation units, cleared with no suppressions. A large
share of them were created by the two preceding commits: making Count() return
base::ssize and the timer durations int64_t moved the mismatch from comparisons
into assignments, which is exactly what this check catches. That is the honest
order for a type migration -- sign-compare first, then truncation.

  EventClass::Frame truncation made explicit, 23 sites   439 -> 405
  _ShapeBufferSize -> int; local counts -> base::ssize    405 -> 323
  strlen() results into int lengths, 68 sites            323 -> 271
  Count()/std::ssize()/time() into int destinations      271 -> 203
  MPlayerMaxAhead and their_frame[] -> int               203 -> 175
  bulk wrap of the remainder, 151 expressions            175 -> 25
  frame cadence, sockets, byte order                      25 -> 0

EventClass::Frame is a 26-bit field in the network packet and the global frame
counter is long, so that truncation is deliberate and now says so. Same for the
packet cadence fields in queue.cc, and for recvfrom/sendto, which return
ssize_t.

HouseClass::Tiberium and Capacity stay long. Narrowing them to int shrank TD's
HouseClass from 3368 to 3360 bytes and td_heap_layout_test failed, which is the
raw-byte save format; the sites got casts instead and the members carry a
comment. That is the second time the layout tests have stopped a member from
being retyped in this tier.

A note on how not to do this. The last 151 sites were wrapped by a script that
took each diagnostic's line and column, walked forward to what it judged the end
of the expression, and inserted a cast around it. The column often points into
the middle of an expression rather than at its start, so this produced about 25
malformed edits -- Array.static_cast<int>(ID(ptr)), x static_cast<int>(- y),
std::max<static_cast<int>(int>(...) -- every one caught by the compiler, and one
that compiled and changed behaviour:

  int graph = kRecordHeight * fixed(kTimerSecond - SpareTicks, kTimerSecond);
  int graph = kRecordHeight * static_cast<int>(fixed(kTimerSecond - SpareTicks), kTimerSecond);

The inserted paren landed on fixed()'s argument separator, turning a two
argument ratio into a one argument construction followed by a comma operator,
so graph became kTimerSecond. It compiled because fixed has a one argument
constructor, and only clang-diagnostic-comma caught it. Two audits over the
other 150 wraps found no second instance: none has a top-level comma inside a
cast, and every changed line reduces to its committed form once the inserted
casts are stripped. Prefer fixing the type over wrapping the expression, and do
not trust a column to delimit an expression.

Fallout fixed here as well: three printf conversions that no longer matched
(MaxAhead and MPlayerMaxAhead became int, a loop index became base::ssize),
three casts left redundant by those same changes, and one uint32_t packet field
compared against the now-signed Session.UniqueID.

Measures 0 tree-wide afterwards, together with clang-diagnostic-comma,
readability-redundant-casting, clang-diagnostic-format, format-signedness,
sign-compare and bugprone-implicit-widening-of-multiplication-result: 0 findings
and no hard error: lines over all 840 units. Both games build and the 84 tests
pass. The full-config sweep was not run for this commit.

+++ src/ra/base.cc
@@ -195,7 +195,7 @@ bool BaseClass::Save(Pipe& file) const {
   */
   file.Put(&House, sizeof(House));
 
-  num_struct = Nodes.Count();
+  num_struct = static_cast<int>(Nodes.Count());
   file.Put(&num_struct, sizeof(num_struct));
 
   /*
@@ -494,7 +494,7 @@ void BaseClass::Write_INI(CCINIClass& ini) {
     **	they must be read in the same order they were created, so "000" must be
     **	read first, etc.
     */
-    ini.Put_Int(INI_Name(), "Count", Nodes.Count());
+    ini.Put_Int(INI_Name(), "Count", static_cast<int>(Nodes.Count()));
 
     // Write each entry into the INI
     for (base::ssize i = 0; i < Nodes.Count(); i++) {
