+++ src/ra/colrlist.cc
@@ -156,7 +156,7 @@ int ColorListClass::Add_Item(int text, RemapControlType* color) {
  *   04/19/1995 BRR : Created.                                             *
  *=========================================================================*/
 void ColorListClass::Remove_Item(const char* text) {
-  int index = List.ID(text);
+  int index = static_cast<int>(List.ID(text));
   if (index != -1) {
     Colors.Delete(index);
     ListClass::Remove_Item(text);
