+++ src/ra/foot.cc
@@ -237,7 +237,7 @@ void FootClass::Debug_Dump(MonoClass* mono) const {
     static const char* _p2c[9] = {"-", "0", "1", "2", "3", "4", "5", "6", "7"};
     for (base::ssize index = 0; index < std::min<base::ssize>(12, std::ssize(Path));
          index++) {
-      mono->Set_Cursor(54 + index, 3);
+      mono->Set_Cursor(static_cast<int>(54 + index), 3);
       mono->Printf("%s",
                    _p2c[((std::abs((int)Path[index] + 1)) % std::ssize(_p2c))]);
     }
@@ -700,7 +700,7 @@ int FootClass::Mission_Guard() {
     }
   }
 
-  return Arm.HasTimeLeft() ? Arm.Value() : dtime + Random_Pick(0, 2);
+  return Arm.HasTimeLeft() ? static_cast<int>(Arm.Value()) : dtime + Random_Pick(0, 2);
 }
 
 /***********************************************************************************************
